Debugged the code by introducing the walkers properly.
Now has to be re-ran to see if I still have a discrepancy between the two ways of computing the correlations...

+++ dimers/ssfsevolve.cpp
@@ -3,11 +3,11 @@
 #include <omp.h>
 using namespace std;
 void ssfsevolve(double J1, int* states, int statesize, int* spinstates, int spinstatesize, int* s2p,
-  int nd,  double* walker2params, double* walker2ids, double* energies, int* failedupdates,
+  int nd,  double* walker2params, int* walker2ids, double* energies, int* failedupdates,
   int nbwalkers, int nthreads, int iters, int nt, int nh) {
     #pragma omp parallel for schedule(dynamic, 1) num_threads(nthreads)
     for(int walker = 0; walker < nbwalkers; walker++) {
-       // for each temperature, do manydualworms niterworm number of times
+       // for each temperature, do  ssfs niterworm number of times
        // get the pointer to the tindex element it state
        int *state = &states[walker*statesize];
        int *spinstate = &spinstates[walker*spinstatesize];
